requirements.txt
asgiref==3.2.3
boto3==1.11.9
botocore==1.14.9
config==0.4.2
Django==3.0.1
django-debug-toolbar==2.1
django-request-logging==0.7.0
django-sslserver==0.22
django-storages==1.8
djangorestframework==3.11.0
docutils==0.15.2
drf-tracking==1.5.0
gunicorn==20.0.4
jmespath==0.9.4
Pillow==7.0.0
psycopg2-binary==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.2
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
